backend/requirements.txt
openai
python-dotenv
fastapi
uvicorn
langgraph
langchain-openai
langchain-core